refactor(ontology): always default-print on stderr
Do not print anything on stdout, because we don't know what the user will want.
Since the `show_ontology_structure` is akin to a log function by default,
it should print on the default log output (i.e. stderr).
Ideally, it would take a `file=sys.stderr` parameter and use its naming scheme when None,
but this is a first step.